Bug 6861: Fast Add returns to circ/circulation.pl in all cases
If Fast Add is accessed from either the circulation homepage or cataloguing,
it should stay on the additem.pl page after saving the first item.  Only when
a Fast Add is done in the middle of a circulation transaction should it redirect
back to circ/circulation.pl to complete that transaction.

This patch adds a conditional check for a borrowernumber input param; the presense
of such implies we're in the middle of a checkout.  It's absense implies we should
remain on additem.pl

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>